fix(spine): ADR-112 slice-B — fold the bot review round (#2261)
All 3 bots (greptile / GCA / CR) reviewed #2261; operator: fix all.

- [greptile-P1 + CR] Attestation staleness: authoringContentHash now folds the
  file-level attestations (splitRef + booleans) into the revision fingerprint, so an
  attestation-only change (e.g. a re-frozen split) appends a fresh row instead of
  reading `unchanged` and keeping the STALE split. authoredAt stays excluded (§8).
- [GCA-high] Trim author/authoredAt/targetDefect/structuralClass at intake
  (.transform before refine) — an untrimmed "alice " bypassed the judgedBy!=author
  guard + minted divergent ids. dslSource left untrimmed (matcher semantics).
- [CR + GCA] Validate authoredAt's CALENDAR at the YAML boundary (reuse
  isIso8601CalendarDate) so a bad date fails CONFIG_INVALID at intake; wrap the
  record-parse in a clean TotemError (the library throws, the command formats —
  declined GCA's set-exitCode-in-library, which violates the gemini-ratified boundary).
- [CR] Deep-freeze the whitelist rows (not just the array).
- [CR] Ledger read-back verifies the row is PRESENT, not last (concurrent-append race).
- [CR-major] Lazy-load: split the synchronous producer into the non-command
  authored-rule-intake lib; commands/rule-author.ts is now a thin lazy-loaded wrapper.
- Tests: attestation-only-edit revision (greptile-P1) + attestation-hash unit.
- DECLINED with evidence (on-thread): CR move-temp-to-.totem/temp — mkdtemp(os.tmpdir())
  is the house pattern in 113 test files; ephemeral + auto-cleaned.

Gate green: pnpm -r build, core 2926 + cli 2839, ESLint core+cli 0 errors, totem lint
PASS, format:check.

Refs strategy#591 (ADR-112 slice B).
